refactoring search module
