+++ src/app/modules/search/search/components/clinic.card/clinic.card.component.scss
