+++ src/app/modules/search/search/components/filter/filter.component.scss
