+++ src/app/modules/search/search/components/paginator/paginator.component.css
